backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.3
psycopg2-binary>=2.9
mysqlclient>=2.2        
PyMySQL>=1.1.1         
python-decouple>=3.8
djangorestframework-simplejwt>=5.3
channels>=4.1
daphne>=4.1
drf-spectacular>=0.27
drf-spectacular-sidecar>=2024.1.1
django-redis>=5.2
django-extensions>=3.2  